debut parsing fait

+++ libft/ft_bzero.c
@@ -13,9 +13,9 @@
 #include "libft.h"
 //#include <stdio.h>
 
-void	ft_bzero(void *s, size_t n)
+void	ft_bzero(void *s, int n)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (i < n)
