+++ libft/ft_memcmp.c
@@ -13,9 +13,9 @@
 //#include <stdio.h>
 #include "libft.h"
 
-int	ft_memcmp(const void *s1, const void *s2, size_t n)
+int	ft_memcmp(const void *s1, const void *s2, int n)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (n > i)
