+++ libft/ft_memmove.c
@@ -13,9 +13,9 @@
 #include "libft.h"
 //#include <stdio.h>
 
-void	*ft_memmove(void *dest, const void *src, size_t n)
+void	*ft_memmove(void *dest, const void *src, int n)
 {
-	size_t	i;
+	int	i;
 
 	if (!dest && !src)
 		return (dest);
