+++ libft/ft_strlen.c
@@ -13,9 +13,9 @@
 #include "libft.h"
 //#include <stdio.h>
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	if (!s)
