+++ libft/ft_strncmp.c
@@ -13,9 +13,9 @@
 //#include <stdio.h>
 #include "libft.h"
 
-int	ft_strncmp(const char *s1, const char *s2, size_t n)
+int	ft_strncmp(const char *s1, const char *s2, int n)
 {
-	size_t			i;
+	int			i;
 
 	i = 0;
 	while (n > i && (((unsigned char *)s1)[i] || ((unsigned char *)s2)[i]))
